Fix plugin tests which selectively apply plugin actions for migrations
We have a number of checks to selectively apply plugin logic. Most
actions are only carried out if we're in the context of a controller
stage or migration action. The anticipated use of stage/migration
annotations has changed since this was originally added to the plugin.
This commit fixes this to work with the current controller code.